requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
paddlepaddle==2.6.2
paddleocr==2.9.1
Pillow==10.4.0
numpy==1.26.4
requests==2.31.0
python-multipart==0.0.6
